requirements.txt
gradio~=3.50
rdkit-pypi
numpy
joblib
python-dotenv